lots of small changes  - null values in array represents missing data  - averge of values into time buckets  - async writing to disk  - db values are all the same size, much less disk used  - range queries with support for '1d' 'now' times  - self reporting of statisitics

+++ libs/compare.c
@@ -9,6 +9,6 @@ typedef struct {
 
 int compare_clocks(const MDB_val *a, const MDB_val *b)
 {
-  return (*(long *)a->mv_data < *(long *)b->mv_data) ? -1 :
-    *(long *)a->mv_data > *(long *)b->mv_data;
+  return (*(long long *)a->mv_data < *(long long *)b->mv_data) ? -1 :
+    *(long long *)a->mv_data > *(long long *)b->mv_data;
 }
